requirements.txt
djangorestframework
rethinkdb
protobuf
requests